[CSL-1970] Fix missing flag(s) causing weird behaviour.
