+++ wallet-new/cardano-sl-wallet-new.cabal
@@ -96,60 +96,13 @@ library
                       ScopedTypeVariables
                       UndecidableInstances
 
---executable wallet-new-server
---  hs-source-dirs:     server
---  main-is:            Main.hs
---  other-modules:      NodeOptions
---  build-depends:
---                        base
---                      , cardano-sl-core
---                      , cardano-sl-infra
---                      , cardano-sl-ssc
---                      , cardano-sl-txp
---                      , cardano-sl-util
---                      , cardano-sl
---                      , cardano-sl-wallet
---                      , formatting
---                      , log-warper
---                      , node-sketch
---                      , optparse-applicative
---                      , time-units
---                      , universum >= 0.1.11
---
---  default-language:    Haskell2010
---  ghc-options:         -threaded -rtsopts
---                       -Wall
---                       -fno-warn-orphans
---                       -O2
---
---  default-extensions:   DeriveDataTypeable
---                        DeriveGeneric
---                        GeneralizedNewtypeDeriving
---                        StandaloneDeriving
---                        FlexibleContexts
---                        FlexibleInstances
---                        MultiParamTypeClasses
---                        FunctionalDependencies
---                        DefaultSignatures
---                        NoImplicitPrelude
---                        OverloadedStrings
---                        RecordWildCards
---                        TypeApplications
---                        TupleSections
---                        ViewPatterns
---                        LambdaCase
---                        MultiWayIf
---                        ConstraintKinds
---                        UndecidableInstances
---                        BangPatterns
---                        TemplateHaskell
---                        GADTs
-
 executable wallet-new-server
   hs-source-dirs:      server
   main-is:             Main.hs
-  other-modules:       NodeOptions
-  ghc-options:         -Wall
+  ghc-options:         -threaded -rtsopts
+                       -Wall
+                       -fno-warn-orphans
+                       -O2
   other-modules:       Cardano.Wallet.API.V0.Handlers
                        Cardano.Wallet.API.V1.Handlers
                        Cardano.Wallet.API.V1.Handlers.Accounts
